+++ src/main/java/com/easystock/repository/CategoryRepository.java
@@ -10,5 +10,5 @@ public interface CategoryRepository extends JpaRepository<Category, Integer> {
 
 	Optional<Category> findByNameIgnoreCase(String name);
 
-	Optional<Category> findById(Long id);
+	Optional<Category> findById(Integer id);
 }
